zamiana double na BigDecimal w CoalConsumptionDTO

+++ src/main/java/pl/ryzykowski/arduino/arduinotemperaturews/config/Globals.java
@@ -3,13 +3,15 @@ package pl.ryzykowski.arduino.arduinotemperaturews.config;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class Globals {
 
     public static int COAL_SUPPLY_MAX_KG;
     public static int COAL_SUPPLY_THRESHOLD_DISTANCE;
     public static double COAL_SUPPLY_COEFFICIENT;
-    public static double COAL_WEIGHT_PER_RATE;
+    public static BigDecimal COAL_WEIGHT_PER_RATE;
 
     @Value("${coal.supply.max.kg}")
     public void setCoalSupplyMaxKg(int coalSupplyMaxKg) {
@@ -23,6 +25,6 @@ public class Globals {
     public void setCoalSupplyCoefficient(double coalSupplyCoefficient) { COAL_SUPPLY_COEFFICIENT = coalSupplyCoefficient; }
 
     @Value("${coal.weight.per.rate}")
-    public void setCoalWeightPerRate(double coalWeightPerRate) {COAL_WEIGHT_PER_RATE = coalWeightPerRate; }
+    public void setCoalWeightPerRate(BigDecimal coalWeightPerRate) {COAL_WEIGHT_PER_RATE = coalWeightPerRate; }
 
 }
